Refactor TopBar component by splitting its functionality into three components: TopBar, PreStamp, and Post.
