+++ frontend/src/components/topBar/TopBar.css
@@ -24,25 +24,6 @@
   flex: 2.5;
 }
 
-.post {
-  width: 70%;
-  height: 40px;
-  background-color: white;
-  color: black;
-  border: none;
-  padding: 10px 20px;
-  font-size: 16px;
-  cursor: pointer;
-  border-radius: 5px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.post:hover {
-  background-color: #cccccc;
-}
-
 .topBarRight {
   flex: 4;
   display: flex;
